+++ src/main/java/zerobase/matching/comment/dto/UpdateComment.java
@@ -11,7 +11,7 @@ public class UpdateComment {
   public static class Request {
 
     @NotNull
-    private long userId;
+    private int userId;
     @NotNull
     private String content;
 
@@ -21,14 +21,14 @@ public class UpdateComment {
   @Getter
   public static class Response {
 
-    private long commentId;
-    private long projectId;
-    private long userId;
+    private int commentId;
+    private int projectId;
+    private int userId;
     private String content;
     private LocalDateTime createTime;
     private LocalDateTime updateTime;
     private LocalDateTime deleteTime;
-    private long parentId;
+    private int parentId;
     private int level;
 
     public static UpdateComment.Response fromEntity(CommentDto commentDto){
